+++ src/nvim/highlight_group.c
@@ -1951,10 +1951,10 @@ static void set_hl_attr(int idx)
   HlAttrs at_en = HLATTRS_INIT;
   HlGroup *sgp = hl_table + idx;
 
-  at_en.cterm_ae_attr = (int16_t)sgp->sg_cterm;
+  at_en.cterm_ae_attr = (int32_t)sgp->sg_cterm;
   at_en.cterm_fg_color = (int16_t)sgp->sg_cterm_fg;
   at_en.cterm_bg_color = (int16_t)sgp->sg_cterm_bg;
-  at_en.rgb_ae_attr = (int16_t)sgp->sg_gui;
+  at_en.rgb_ae_attr = (int32_t)sgp->sg_gui;
   // FIXME(tarruda): The "unset value" for rgb is -1, but since hlgroup is
   // initialized with 0 (by garray functions), check for sg_rgb_{f,b}g_name
   // before setting attr_entry->{f,g}g_color to a other than -1
